idl - Testing and debugging now works buildSrc.
