+++ buildSrc/build.gradle.kts
@@ -9,4 +9,5 @@ repositories {
 dependencies {
     implementation("org.jetbrains.kotlin:kotlin-reflect")
     implementation("com.squareup:kotlinpoet:1.5.0")
+    implementation(kotlin("test-junit"))
 }
